pre-processing/requirements.txt
fastapi
uvicorn
numpy
requests
psycopg2-binary
meshio
h5py
scipy